requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
Django==3.1.5
docx2pdf==0.1.7
fpdf==1.7.2
future==0.18.2
html5lib==1.1
img2pdf==0.4.0
lxml==4.6.2
pdf2image==1.14.0
pdfkit==0.6.1
pikepdf==2.4.0
Pillow==8.1.0
PyPDF2==1.26.0
pypiwin32==223
python-bidi==0.4.2
pytz==2020.5
pywin32==227
reportlab==3.5.60
six==1.15.0
sqlparse==0.4.1
tqdm==4.56.0
webencodings==0.5.1
xhtml2pdf==0.2.5
